refactor distribution names
