+++ NAMESPACE
@@ -15,9 +15,9 @@ export(Parameter)
 export(ParameterVector)
 export(Population)
 export(SetFIMSFunctions)
-export(TMBDlnormDistribution)
-export(TMBDmultinomDistribution)
-export(TMBDnormDistribution)
+export(DlnormDistribution)
+export(DmultinomDistribution)
+export(DnormDistribution)
 export(ToJSON)
 export(clear)
 export(clear_logs)
